amended the routes for refactored code

+++ dev/setup/config.js
@@ -1,10 +1,22 @@
 'use strict';
 
 import {configure} from '../utils'
-const express = require('express');
-import eJwt from 'express-jwt';
+const express = require('express')
+import eJwt from 'express-jwt'
 import {slugify} from '../utils'
 import keys from '../../secrets.json'
+
+import multer from 'multer'
+const storage = multer.diskStorage({
+    destination: function (req, file, cb) {
+        cb(null, './dev/uploads')
+    },
+    filename: function (req, file, cb) {
+        cb(null, file.originalname)
+  }
+})
+
+const upload = multer({ storage: storage })
 
 export let config = configure({
   api: express.Router(),
@@ -12,16 +24,18 @@ export let config = configure({
 })
 
 
-const addContent = function(req,res,next,content){
+export const addContent = function(args,req,res,next){
+  const state = args
+  console.log(state)
   if(req.method !== "POST"){
-    let data = Object.assign(content, req.body);
+    var data = Object.assign(state.content, req.body);
   }else{
-    let data = new this.db(content);
+    var data = new state.model(state.content);
   }
 
   data.saveAll()
   .then((result) => {
-    return res.send({"message": `document added to ${this.db.getTableName()}`});
+    return res.send({"message": `document added to ${state.model.getTableName()}`});
   })
   .error((error) => {
     res.json({"message": error.toString()});
@@ -36,47 +50,54 @@ export const getData = (args) => {
     all(){
       return state.model.run();
     },
-    single(slug){
-      return state.model.filter({slug: slug})
+    single(query){
+      console.log(query)
+      return state.model.filter(query)
     }
   }
 }
 
 const resource = (state) => ({
-    get(req,res,next) {
+    get: (req,res,next) => {
       getData(state).all().then((results) => {
         res.json(results);
       })
     },
     getSingle: (req,res,next) => {
-      getData(state).single(req.params.slug)
+      getData(state).single({slug: req.params.slug})
       .then((results) => {
         res.json(results[0]);
       })
     },
     post: (req,res,next) => {
-      let slug = slugify(req.body.title);
-      getData(state).single(slug)
+      let slug = slugify(req.body.title)
+      getData(state).single({slug: slug})
       .then((results) =>{
         if(results.length === 0){
-          addContent(req,res,next,req.body);
+          addContent({
+            model: state.model,
+            content: req.body
+          },req,res,next);
         }else{
           res.json({'message': 'This item already exists'})
         }
       });
     },
     put: (req,res,next) => {
-      getData(state).single(req.params.slug)
+      getData(state).single({slug: req.params.slug})
       .then((results) => {
         if(results.length !== 0){
-          addContent(req,res,next,results[0])
+          addContent({
+            model: state.model,
+            content: results[0]
+          },req,res,next);
         }else{
           res.json({'message': 'There is no document with this slug'})
         }
       })
     },
     delete: (req,res,next) => {
-      getData(state).single(req.params.slug)
+      getData(state).single({slug: req.params.slug})
       .delete()
       .then((results) => res.json({'message': 'item deleted'}))
     }
@@ -91,39 +112,69 @@ const secureRoute = function(secure){
   }
 }
 
-export const registerRoute = function(config, routes){
+const removeMethods = (state) => ({
+  removeMethods: (methods) => {
+    methods.map((item) => {
+      delete state[item.toLowerCase()]
+    })
+  }
+})
+
+export const resourceRoutes = function(config, routes){
   const routess = Object.keys(routes)
   routess.map((item) => {
     let route = routes[item]
 
-    config.api.get(
-      `/${route.route}`,
-      secureRoute(route.secure.get),
-      route.get
-    )
+    if(route.get !== undefined){
+      config.api.get(
+        `/${route.route}`,
+        secureRoute(route.secure.get),
+        route.get
+      )
+    }
 
-    config.api.get(
-      `/${route.route}/:slug`,
-      secureRoute(route.secure.get),
-      route.getSingle
-    )
+    if(route.get !== undefined){
+      config.api.get(
+        `/${route.route}/:slug`,
+        secureRoute(route.secure.get),
+        route.getSingle
+      )
+    }
 
-    config.api.put(
-      `/${route.route}/:slug`,
-      secureRoute(route.secure.put),
-      route.put
-    )
+    if(route.put !== undefined){
+      config.api.put(
+        `/${route.route}/:slug`,
+        secureRoute(route.secure.put),
+        route.put
+      )
+    }
 
-    config.api.delete(
-      `/${route.route}/:slug`,
-      secureRoute(route.secure.delete),
-      route.delete
-    )
+    if(route.delete !== undefined){
+      config.api.delete(
+        `/${route.route}/:slug`,
+        secureRoute(route.secure.delete),
+        route.delete
+      )
+    }
 
-    config.api.post(
-      `/${route.route}/`,
-      secureRoute(route.secure.post),
-      route.post)
+    if(route.post !== undefined){
+      config.api.post(
+        `/${route.route}/`,
+        secureRoute(route.secure.post),
+        route.post
+      )
+    }
+
+    if(route.filePost !== undefined){
+      config.api.post(
+        `/${route.route}/`,
+        secureRoute(route.secure.post),
+        upload.array('file'),
+        route.filePost
+      )
+    }
+
+
   })
 }
 
@@ -132,14 +183,17 @@ export const createRoute = function(args){
     route: args.route,
     model: args.model,
     secure: {
-      get: args.secure.get || false,
-      post: args.secure.post || true,
-      put: args.secure.put || true,
-      delete: args.secure.delete || true,
+      get: false,
+      post: true,
+      put: true,
+      delete: true,
     }
   }
   return Object.assign(
     state,
-    resource(state)
+    args,
+    resource(state),
+    removeMethods(state)
   )
+
 }
